Move shell-specific script generation into adapters
